requirements-dopc.txt
absl-py==2.3.1
accelerate==1.6.0
adaptive-boxes==0.0.4
aim==3.29.1
aim-ui==3.29.1
aimrecords==0.0.7
aimrocks==0.5.2
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.15.2
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.13.0
asttokens==3.0.1
attrs==25.3.0
autopage==0.6.0
base58==2.0.1
blinker==1.9.0
boto3==1.42.81
botocore==1.42.81
cachetools==6.0.0
certifi==2025.4.26
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.2
click==8.2.1
cliff==4.13.2
cloudpickle==3.1.1
cmaes==0.13.0
cmd2==3.4.0
colorlog==6.9.0
colour==0.1.5
contourpy==1.3.2
cryptography==46.0.5
cubinlinker-cu11==0.3.0.post3
cuda-bindings==11.8.7
cuda-python==11.8.7
cudf-cu11==25.6.0
cuml-cu11==25.6.0
cupy-cuda11x==13.4.1
cuvs-cu11==25.6.0
cycler==0.12.1
dask==2025.5.0
dask-cuda==25.6.0
dask-cudf-cu11==25.6.0
datasets==3.6.0
datatable==1.1.0
decorator==5.2.1
dill==0.3.8
distlib==0.4.0
distributed==2025.5.0
distributed-ucxx-cu11==0.44.0
docopt==0.6.2
docstring_parser==0.17.0
einops==0.8.1
et_xmlfile==2.0.0
evaluate==0.4.3
executing==2.2.1
fastapi==0.135.3
fastrlock==0.8.3
filelock==3.25.2
fonttools==4.58.0
frozenlist==1.6.0
fsspec==2025.3.0
gdspy==1.6.13
gdstk==0.9.62
gitdb==4.0.12
GitPython==3.1.46
glop==0.1.1
greenlet==3.2.2
grpcio==1.76.0
gsplat==1.5.3
h11==0.16.0
huggingface==0.0.1
huggingface-hub==0.30.2
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
identify==2.6.18
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
iniconfig==2.3.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
jaxtyping==0.3.3
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.1
jsonpickle==4.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.8
klayout==0.28.17.post1
layout2svg==0.1.6
libcudf-cu11==25.6.0
libcuml-cu11==25.6.0
libcuvs-cu11==25.6.0
libkvikio-cu11==25.6.0
libraft-cu11==25.6.0
librmm-cu11==25.6.0
libucx-cu11==1.18.1
libucxx-cu11==0.44.0
lightning==2.6.1
lightning-utilities==0.15.3
llvmlite==0.44.0
locket==1.0.0
loguru==0.7.3
lxml==4.9.4
Mako==1.3.10
mapbox_earcut==1.0.3
Markdown==3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
multiprocess==0.70.16
narwhals==2.18.0
networkx==3.4.2
ninja==1.11.1.4
nodeenv==1.10.0
numba==0.61.2
numba-cuda==0.11.0
numpy==1.26.4
nvdiffrast @ file:///home/xiaye/dac2025/splat-opc/nvdiffrast
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
nvtx==0.2.12
omegaconf==2.3.0
opencv-python==4.11.0.86
openpyxl==3.1.5
optuna==2.10.1
packaging==25.0
pandas==2.2.3
parso==0.8.6
partd==1.4.2
pdfminer.six==20251230
pdfplumber==0.11.9
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.9.4
plotly==6.6.0
pluggy==1.6.0
plyfile==1.1
praat-parselmouth==0.4.6
pre_commit==4.5.1
prettytable==3.17.0
prompt_toolkit==3.0.52
propcache==0.3.1
protobuf==6.33.0
psutil==7.0.0
ptxcompiler-cu11==0.8.1.post3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pycommon==0.3.6
pycparser==2.22
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.1
pylibcudf-cu11==25.6.0
pylibraft-cu11==25.6.0
pynndescent==0.6.0
pynvml==12.0.0
pyparsing==3.2.3
pypdfium2==5.6.0
pyperclip==1.11.0
pytest==9.0.2
python-dateutil==2.9.0.post0
python-discovery==1.2.1
python-dotenv==1.2.2
python-gdsii==0.2.1
pytorch-lightning==2.6.1
pytz==2025.2
pyvis==0.3.2
pyworld==0.3.5
PyYAML==6.0.2
raft-dask-cu11==25.6.0
rapids-dask-dependency==25.6.0
rapids-logger==0.1.1
referencing==0.37.0
regex==2024.11.6
requests==2.32.3
RestrictedPython==8.1
rich==14.3.3
rich-argparse==1.7.2
rmm-cu11==25.6.0
rootutils==1.0.7
rpds-py==0.30.0
s3transfer==0.16.0
safetensors==0.5.3
scikit-learn==1.8.0
scipy==1.15.3
sentence-transformers==5.3.0
shtab==1.7.2
six==1.17.0
smmap==5.0.3
sortedcontainers==2.4.0
soundfile==0.13.1
spark-parser==1.8.9
SQLAlchemy==2.0.40
stack-data==0.6.3
starlette==1.0.0
stevedore==5.7.0
streamlit==1.55.0
sympy==1.14.0
tblib==3.1.0
tenacity==9.1.4
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
timm==1.0.25
tokenizers==0.21.1
toml==0.10.2
toolz==1.0.0
torch==2.7.0
torchaudio==2.7.0
torchmetrics==1.9.0
torchsummary==1.5.1
torchvision==0.22.0
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.52.3
treelite==4.4.1
triton==3.3.0
typeguard==4.4.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tyro==0.9.35
tzdata==2025.2
ucx-py-cu11==0.44.0
ucxx-cu11==0.44.0
umap-learn==0.5.11
uncompyle==2.0.0
uncompyle6==3.9.2
urllib3==2.4.0
uvicorn==0.42.0
virtualenv==21.2.0
wadler_lindig==0.1.7
watchdog==6.0.0
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.3
xdis==6.1.4
xxhash==3.5.0
yarl==1.20.0
zict==3.0.0
zipp==3.22.0

